Initial commit
This commit lays a (semi-functional) foundation, that includes:

- an Openire server (openfire01), which uses an embedded database.
- a nameserver (bind01), which provides the relevant DNS records.
- a Kerberos Key Distribution Service, which doesn't do anything yet.

This is by no means a finished product at this point. Although Openfire
will start, there are no users other than the default admin user
(admin/admin) and Kerberos is not hooked in at all. There is probably
a lot of room for improvement on the Vagrant and Ansible side of things
too.